iter6: fixs

+++ migrations/20250826181206_add_table_draw_history.sql
@@ -4,7 +4,7 @@ CREATE TABLE IF NOT EXISTS user_draw_history
 (
     user_id        int NOT NULL,
     transaction_dt TIMESTAMP,
-    order_id       int NOT NULL,
+    order_id       bigint NOT NULL,
     draw           int
 );
